+++ testing/simulator/runner/cli.rs
@@ -151,6 +151,12 @@ pub struct SimulatorCLI {
         default_value_t = false
     )]
     pub disable_savepoint_rollback: bool,
+    #[clap(
+        long,
+        help = "disable Overflow-Chain-Integrity Property",
+        default_value_t = false
+    )]
+    pub disable_overflow_chain_integrity: bool,
     #[clap(long, help = "disable FsyncNoWait Property", default_value_t = true)]
     pub disable_fsync_no_wait: bool,
     #[clap(long, help = "disable FaultyQuery Property")]