+++ testing/simulator/runner/env.rs
@@ -1105,6 +1105,7 @@ impl SimulatorEnv {
             disable_union_all_preserves_cardinality: cli_opts
                 .disable_union_all_preserves_cardinality,
             disable_savepoint_rollback: cli_opts.disable_savepoint_rollback,
+            disable_overflow_chain_integrity: cli_opts.disable_overflow_chain_integrity,
             disable_fsync_no_wait: cli_opts.disable_fsync_no_wait,
             disable_faulty_query: cli_opts.disable_faulty_query,
             page_size: 4096, // TODO: randomize this too
@@ -1443,6 +1444,7 @@ pub(crate) struct SimulatorOpts {
     pub(crate) disable_where_true_false_null: bool,
     pub(crate) disable_union_all_preserves_cardinality: bool,
     pub(crate) disable_savepoint_rollback: bool,
+    pub(crate) disable_overflow_chain_integrity: bool,
     pub(crate) disable_fsync_no_wait: bool,
     pub(crate) disable_faulty_query: bool,
     pub(crate) disable_reopen_database: bool,